General code cleanup.
o removed old stuff no longer used
o fixed some "eclipse" warnings
o fixed a lot of autoboxing/casting stuff
o used FindBugs and actually found 1 real bug 
  (XaLogicalLog leaked in a hashmap for read only tx's during recovery)


git-svn-id: https://svn.neo4j.org/projects/neo/trunk@59 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ enterprise/src/java/org/neo4j/impl/nioneo/store/PropertyRecord.java
@@ -128,7 +128,7 @@ public class PropertyRecord
 
 	public void clearKeyRecords()
 	{
-		int nextBlockId = ( int ) keyBlock;
+		int nextBlockId = keyBlock;
 		while ( nextBlockId != Record.NO_NEXT_BLOCK.intValue() ) 
 		{
 			DynamicRecord record = keyRecords.get( nextBlockId );
